Spec Validation Against AWS Cognito - maximum length is 20 characters
While creating a User Pool in AWS Cognito, we can only set minimum characters and there is no option to set maximum characters in AWS Cognito.
Existing Spec: maximum length is 20 characters
Request for change: 1. Can we remove this spec as we don't have the option to set it in AWS Cognito? 2. We can do this validation from front-end as soon as User clicks on Register/Sign Up.